+++ CHANGES.md
@@ -4,7 +4,8 @@
 - added type test, if parameter gets checked for "all"
 - specify that secrets_weatherDB file should be on PYTHONPATH environment variable
 - Changed DGM5 to Copernicus DGM25, because of license advantages
-- 
+- add kwargs to update_richter_class of StationsN
+
 ## Version 0.0.13
 - change the timezone allocation method of the precipitation download df
 - set freq to 10 minutes of precipitation download, to be able to overwrite Values with NAs